Remove PITA stats.
Tried to track a lot of different timing stats, but all of the little
bits of overhead that weren't getting tracked at each node were adding
up until large (percentage) chunks of the overall test run were
unknowable. It isn't very useful to tell someone that 5% of the run
was spent in their hooks, 20% in their tests, 25% in the measurable
parts of anaphora code, if you have nothing meaningful to say about
where the other 50% is going.